pyproject.toml
[tool.poetry]
name = "invisiblewatermark"
version = "0.1.0"
description = ""
authors = ["lijianan <lijianan@tezign.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
numpy = "^1.25.2"
pillow = "^10.0.0"
typer = "^0.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
ivwm = "invisiblewatermark.cli:app"

